+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 ### Features
 
+* make thread_count configurable for parallel_bulk in opensearch output connector
+
 ### Improvements
 
 ### Bugfix